Move BinaryTupleContainer to binarytuple module
