+++ modules/binary-tuple/src/main/java/org/apache/ignite/internal/binarytuple/BinaryTupleContainer.java
@@ -15,13 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.schema;
+package org.apache.ignite.internal.binarytuple;
 
-import org.apache.ignite.internal.binarytuple.BinaryTupleReader;
 import org.jetbrains.annotations.Nullable;
 
 /**
- * An entity that stores underlying data in a {@link BinaryTuple}.
+ * An entity that stores underlying data in a {@link BinaryTupleReader}.
  */
 @SuppressWarnings("InterfaceMayBeAnnotatedFunctional")
 public interface BinaryTupleContainer {
